Add UNICAST_MASTER_TABLE_NP management TLV
This change adds new UNICAST_MASTER_TABLE_NP management TLV that
exposes table with details of each gm: identity, IP address,
state, if it's selected as best master, and information from
the latest Announce msg that is used in BMCA algorithm:
ClockQuality and Priorities.

The inspiration was `chronyc sources` command that provides
great insights into the state of NTP client. Here we tried
to provide something similar, but for unicast PTP client.

Justification: we operate unicast PTPv2 in datacenters,
and each client has multiple unicast GMs set up in the config.

First, it quickly became clear that we need a way to map the
GM identity from PARENT_DATA_SET or TIME_STATUS_NP TLVs
to the actual IP.

Next we realized the need to be able to inspect the full
unicast master table along with GM states and preferably
Announce information, to quickly debug potential issues
and better understand why the particular GM was selected
as best master.

The 1588-2019 standard doesn't provide any real means to inspect the
active unicast master table. UNICAST_MASTER_TABLE TLV is pretty
useless as described in the standard, and with linuxptp it can
be effectively replaced by `cat /etc/ptp4l.conf`.

Here is the example output that hopefully shows the
usefullness of new TLV: we can instantly see that
ptp4l coudn't even talk to some of the GMs.

  ./pmc -u -b 0  'GET UNICAST_MASTER_TABLE_NP'
  sending: GET UNICAST_MASTER_TABLE_NP
	  4857dd.fffe.0e91da-1 seq 0 RESPONSE MANAGEMENT MID_UNICAST_MASTER_TABLE_NP
		  actual_table_size 9
		  BM  identity                 address                            state     clockClass clockQuality offsetScaledLogVariance p1  p2
		  no  b8cef6.fffe.7349d4-1     2401:db00:2515:f001:face:0:2a3:0   HAVE_ANN  6          0x21         0x59e0                  128 128
		  yes b8cef6.fffe.0210e4-1     2401:db00:2515:f001:face:0:3d1:0   HAVE_SYDY 6          0x21         0x59e0                  128 128
		  no  b8cef6.fffe.057e20-1     2401:db00:2515:f001:face:0:3fa:0   HAVE_ANN  6          0x21         0x59e0                  128 128
		  no  b8cef6.fffe.7349dc-1     2401:db00:2515:f001:face:0:da:0    HAVE_ANN  6          0x21         0x59e0                  128 128
		  no  ffffff.ffff.ffffff-65535 2401:db00:2515:f002:face:0:11b:0   WAIT      0          0x00         0x0000                  0   0
		  no  b8cef6.fffe.7349c4-1     2401:db00:2515:f002:face:0:1ec:0   HAVE_ANN  6          0x21         0x59e0                  128 128
		  no  ffffff.ffff.ffffff-65535 2401:db00:2515:f002:face:0:94:0    WAIT      0          0x00         0x0000                  0   0
		  no  ffffff.ffff.ffffff-65535 192.168.0.1                        WAIT      0          0x00         0x0000                  0   0
		  no  b8cef6.fffe.7349c8-1     2401:db00:2515:f002:face:0:b7:0    HAVE_ANN  6          0x21         0x59e0                  128 128

I realize that this is can be controversial
(for one, no other PMC command provides table-like output),
so I'm happy to hear your suggestion.

Signed-off-by: Alexander Bulimov <abulimov@fb.com>